+++ packages/virtual/core/src/index.server.ts
@@ -1,0 +1,2 @@
+import "./resize-observer-node";
+export * from "./index";