+++ packages/virtual/core/src/resize-observer-node.ts
@@ -1,0 +1,5 @@
+global.ResizeObserver ||= class implements ResizeObserver {
+    observe() {}
+    unobserve() {}
+    disconnect() {}
+};